# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BRAVE Program Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/review-queue:
    get:
      operationId: getAdminReviewQueue
      tags: [admin]
      summary: Get all pending financial entries for review
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [revenue]
        - name: campusId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Review queue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminReviewQueueResponse"

  # Leaderboard
...
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [student, coordinator, admin, superadmin]
        - name: campusId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [admin]
      summary: Create a coordinator or admin account (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/users/{id}:
    patch:
      operationId: updateUser
      tags: [admin]
      summary: Update a user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # Programme config
  /admin/programme-config:
    get:
      operationId: getProgrammeConfig
      tags: [admin]
      summary: Get programme configuration
      responses:
        "200":
          description: Programme config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgrammeConfig"
    patch:
      operationId: updateProgrammeConfig
      tags: [admin]
      summary: Update programme configuration (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProgrammeConfigBody"
      responses:
        "200":
          description: Updated config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgrammeConfig"

  # Audit log
  /admin/audit-log:
    get:
      operationId: getAuditLog
      tags: [admin]
      summary: Get audit log (admin)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Audit log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLogEntry"

  # Roster management
  /admin/roster:
    get:
      operationId: listRosterEntries
      tags: [roster]
      summary: List roster entries
      parameters:
        - name: campusId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Roster entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RosterEntry"
    post:
      operationId: addRosterEntry
      tags: [roster]
      summary: Add a student to the roster
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddRosterEntryBody"
      responses:
        "201":
          description: Roster entry added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RosterEntry"
  /admin/roster/import:
    post:
      operationId: bulkImportRoster
      tags: [roster]
      summary: Bulk import students into roster from parsed Excel data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkImportRosterBody"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkImportRosterResponse"
  /admin/access-requests:
    get:
      operationId: listAccessRequests
      tags: [roster]
      summary: List access requests
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Access requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccessRequest"
  /admin/access-requests/{id}:
    patch:
      operationId: updateAccessRequest
      tags: [roster]
      summary: Approve or reject an access request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccessRequestBody"
      responses:
        "200":
          description: Updated access request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessRequest"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        replitId:
          type: ["string", "null"]
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        profileImage:
          type: ["string", "null"]
        role:
          type: string
          enum: [student, coordinator, admin, superadmin]
        campusId:
          type: ["integer", "null"]
        campusName:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - firstName
        - lastName
        - role
        - isActive
        - createdAt
    CreateUserBody:
      type: object
      properties:
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [coordinator, admin]
        campusId:
          type: ["integer", "null"]
        password:
          type: string
        formsUserId:
          type: ["string", "null"]
          description: Optional Forms SSO user_id used to link this account to its first SSO sign-in.
      required:
        - email
        - firstName
        - lastName
        - role
        - password
    UpdateUserBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [student, coordinator, admin]
        campusId:
          type: ["integer", "null"]
        isActive:
          type: boolean
...
    AdminReviewQueueItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [revenue]
        teamId:
          type: integer
        teamName:
          type: string
        campusName:
          type: string
        projectTitle:
          type: string
        clientName:
          type: string
        amount:
          type: integer
        submittedAt:
          type: string
          format: date-time
        isOverdue:
          type: boolean
        supportingDocUrl:
          type: ["string", "null"]
        paymentProofUrl:
          type: ["string", "null"]
        invoiceUrl:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - id
        - type
        - teamId
        - teamName
        - campusName
        - projectTitle
        - clientName
        - amount
        - submittedAt
        - isOverdue
    AdminReviewQueueResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminReviewQueueItem"
        overdueCount:
          type: integer
        totalCount:
          type: integer
      required:
        - items
        - overdueCount
        - totalCount
...
    ProgrammeConfig:
      type: object
      properties:
        id:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        demoDayDate:
          type: ["string", "null"]
          format: date
        demoEligibilityThreshold:
          type: integer
        leaderboardFrozen:
          type: boolean
        demoDayApplicationsOpen:
          type: boolean
        demoDayApplicationDeadline:
          type: ["string", "null"]
          format: date
        programmePhase:
          type: string
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - startDate
        - endDate
        - demoEligibilityThreshold
        - leaderboardFrozen
        - demoDayApplicationsOpen
        - programmePhase
        - updatedAt
    UpdateProgrammeConfigBody:
      type: object
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        demoDayDate:
          type: ["string", "null"]
          format: date
        demoEligibilityThreshold:
          type: integer
        leaderboardFrozen:
          type: boolean
        demoDayApplicationsOpen:
          type: boolean
        demoDayApplicationDeadline:
          type: ["string", "null"]
          format: date
        programmePhase:
          type: string
    AuditLogEntry:
      type: object
      properties:
        id:
          type: integer
        actorId:
          type: integer
        actorName:
          type: string
        action:
          type: string
        targetType:
          type: string
        targetId:
          type: ["integer", "null"]
        details:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - actorId
        - actorName
        - action
        - targetType
        - createdAt
    RosterEntry:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: string
        fullName:
          type: string
        email:
          type: string
        campusName:
          type: string
        campusId:
          type: ["integer", "null"]
        niatId:
          type: ["string", "null"]
        batchSectionName:
          type: ["string", "null"]
        isWhitelisted:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - studentId
        - fullName
        - email
        - campusName
        - isWhitelisted
        - createdAt
    AddRosterEntryBody:
      type: object
      properties:
        studentId:
          type: string
        fullName:
          type: string
        email:
          type: string
        campusName:
          type: string
        campusId:
          type: ["integer", "null"]
        niatId:
          type: ["string", "null"]
        batchSectionName:
          type: ["string", "null"]
        isWhitelisted:
          type: boolean
      required:
        - studentId
        - fullName
        - email
        - campusName
    BulkImportRosterBody:
      type: object
      properties:
        students:
          type: array
          items:
            type: object
            properties:
              studentUserId:
                type: string
              studentName:
                type: string
              niatId:
                type: string
              instituteName:
                type: string
              batchSectionName:
                type: string
            required:
              - studentName
              - instituteName
      required:
        - students
    BulkImportRosterResponse:
      type: object
      properties:
        inserted:
          type: integer
        skipped:
          type: integer
        total:
          type: integer
      required:
        - inserted
        - skipped
        - total
    AccessRequest:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        batch:
          type: ["string", "null"]
        niatId:
          type: ["string", "null"]
        campusName:
          type: string
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - fullName
        - email
        - campusName
        - status
        - createdAt
...
    UpdateAccessRequestBody:
      type: object
      properties:
        status:
          type: string
          enum: [approved, rejected]
        notes:
          type: string
      required:
        - status
...