# lib/api-zod/src/generated/types/createUserBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BRAVE Program Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateUserBodyRole } from "./createUserBodyRole";

export interface CreateUserBody {
  email: string;
  firstName: string;
  lastName: string;
  role: CreateUserBodyRole;
  /** @nullable */
  campusId?: number | null;
  password: string;
  /**
   * Optional Forms SSO user_id used to link this account to its first SSO sign-in.
   * @nullable
   */
  formsUserId?: string | null;
}